refactor: move EmailDeliveryStatusCallback under AccreditedRepresentativePortal namespace
