+++ modules/accredited_representative_portal/spec/services/accredited_representative_portal/email_delivery_status_callback_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe EmailDeliveryStatusCallback do
+RSpec.describe AccreditedRepresentativePortal::EmailDeliveryStatusCallback do
   let(:base_metadata) do
     {
       'form_number' => 'Form Number',
